refactor: extract app_helpers from api/app.py
